refactor: :recycle: change sign_in_mod references to signIn
