+++ mintapi/cli.py
@@ -10,7 +10,7 @@ import keyring
 import configargparse
 
 from mintapi.api import Mint
-from mintapi.sign_in_mod import get_email_code
+from mintapi.signIn import get_email_code
 
 logger = logging.getLogger("mintapi")
